requirements.txt
asgiref==3.2.10
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-crispy-forms==1.9.2
django-environ==0.4.5
django-import-export==2.3.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
et-xmlfile==1.0.1
google-api-core==1.27.0
google-auth==1.30.0
google-cloud-aiplatform==0.9.0
google-cloud-bigquery==2.16.1
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
grpcio==1.37.1
gunicorn==20.1.0
idna==2.10
jdcal==1.4.1
joblib==1.0.1
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.9
pandas==1.2.4
plotly==4.14.3
proto-plus==1.18.1
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.23.0
scipy==1.6.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
threadpoolctl==2.1.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
